Revert "Ei 162 create draft incentive application"

+++ src/SFA.DAS.EmployerIncentives.Database/Tables/IncentiveApplication.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[IncentiveApplication]
 (
 	[Id] UNIQUEIDENTIFIER NOT NULL PRIMARY KEY,
-	[AccountId] BIGINT NOT NULL,
-	[AccountLegalEntityId] BIGINT NOT NULL,
+	[AccountId] INT NOT NULL,
+	[AccountLegalEntityId] INT NOT NULL,
 	[DateCreated] DATETIME NOT NULL,
 	[Status] NVARCHAR(50) NOT NULL,
 	[DateSubmitted] DATETIME2 NULL,
